Version 1.1
Added a "-q rows,cols" option to mult which makes it possible
to do computations in the quantum cohomology ring of Gr(d,n)
where d=rows and n=rows+cols.
The maple interface has new function "quantum(rows,cols)" which
causes tos to calculate in this ring as well.