Version 1.1.6 (sage)
This is the first release of a branch of LRCALC aimed to make it
compatible with SAGE.  The GNU automake/autoconf build system is
now used, and the Schubert code from schmult-1.1.5.tar.gz has
been incorporated.  Special thanks to Nicolas M. Thiéry and
Jean-Pierre Flori for their help with this release.  See the
ChangeLog file for more details.