Version 1.1.7
The lrcalc library now uses setjmp() and longjmp() to recover from
"out of memory" events.  In addition, the command "lrcalc" has been
added, which is able to perform the functions mult, skew, coprod,
lrcoef, and lrskew.  The purpose of these changes is to make the LR
Calculator more compatible with SAGE and Fedora.  Finally, the program
"schubmult" for multiplication of Schubert polynomials has been added.