Version 1.2
This release combines the two branches, so that the lrcalc-sage branch
is no longer needed.

The directory structure is changed.

The GNU automake/autoconf build system is used to build the programs.
Run "autoreconf -i" to generate the configure script.

The programs mult, skew, etc. are replaced with a single program
lrcalc.  This will put fewer files in /usr/local/bin.